# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: RemoteGPU API
  description: RemoteGPU 远程GPU管理平台 API 文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@remotegpu.com
servers:
  - url: http://0.0.0.:8080/api/v1
    description: 开发环境
  - url: https://0.0.0.0:8080/api/v1
    description: 生产环境
...
security:
  - BearerAuth: []
paths:
...
  /admin/dashboard/stats:
    get:
      tags:
        - Admin - Dashboard
      summary: 获取仪表板统计数据
      description: 获取系统整体统计信息
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/alerts:
    get:
      tags:
        - Admin - Monitoring
      summary: 获取告警列表
      description: 获取系统告警信息
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: level
          in: query
          schema:
            type: string
            enum: [info, warning, error, critical]
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  total:
                    type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 认证令牌
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 错误码（例如：2003=密码错误，2006=账号禁用）
          example: 2003
        msg:
          type: string
          description: 错误信息
          example: 密码错误
        data:
          nullable: true
          description: 可选错误数据
...
    DashboardStats:
      type: object
      properties:
        total_machines:
          type: integer
        available_machines:
          type: integer
        allocated_machines:
          type: integer
        total_customers:
          type: integer
        active_tasks:
          type: integer
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        level:
          type: string
          enum: [info, warning, error, critical]
        message:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
...